refactor: resolve dev server builds through Effect without a Promise round-trip
startServerBuildEvaluationEffect wrapped the Promise-based
resolveServerBuildModule, which itself runs an Effect, so every dev server
build evaluation crossed the Effect/Promise boundary twice. The Effect-based
resolution now lives in an internal server-build-resolution module that
dev-runtime-artifacts composes directly; server-utils keeps the public
Promise wrappers so the published declaration graph stays free of effect
types (re-verified by walking dist/index.d.ts imports).
